packet read/write refactored

+++ common/src/main/java/hu/lakati/ihome/common/event/DataEvent.java
@@ -6,5 +6,5 @@ import lombok.experimental.SuperBuilder;
 @Getter
 @SuperBuilder
 public class DataEvent extends BaseEvent {
-    private final int value;
+    private final double value;
 }
